requirements.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.5
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
PyNaCl==1.5.0
Werkzeug==2.3.6
